refactor: Extract runOptimizeScan, runCostsScan, runCloudCostsScan, buildCostBreakdownFromScenarios, deriveCostConfidence to costs.go
