+++ cmd/opscart-scan/emergency.go
@@ -1,0 +1,23 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/opscart/opscart-k8s-watcher/pkg/scanner"
+)
+
+func runEmergencyScan(clusterContext string) error {
+	fmt.Printf("\n🔍 Cluster: %s\n", clusterContext)
+	s, err := scanner.NewScanner(clusterContext)
+	if err != nil {
+		return fmt.Errorf("connecting to cluster: %w", err)
+	}
+
+	issues, err := s.FindEmergencyIssues(namespace)
+	if err != nil {
+		return fmt.Errorf("scanning cluster: %w", err)
+	}
+
+	scanner.PrintEmergencyIssues(issues)
+	return nil
+}
