+++ cmd/opscart-scan/resources.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/opscart/opscart-k8s-watcher/pkg/analyzer"
+)
+
+func runResourcesScan(clusterContext string) error {
+	fmt.Printf("\n🔍 Cluster: %s\n", clusterContext)
+	clientset, err := getKubernetesClient(clusterContext)
+	if err != nil {
+		return fmt.Errorf("connecting to cluster: %w", err)
+	}
+
+	ra := analyzer.NewResourceAnalyzer(clientset)
+	analysis, err := ra.AnalyzeClusterResources(namespace)
+	if err != nil {
+		return fmt.Errorf("analyzing resources: %w", err)
+	}
+
+	analyzer.PrintResourceAnalysis(analysis, format)
+	return nil
+}
